feat: change the table transaction field amount type changed to int type

+++ src/database/database.ts
@@ -25,7 +25,7 @@ const createTables = async () => {
         id SERIAL PRIMARY KEY,
         payer INT NOT NULL,
         payee INT NOT NULL,
-        amount NUMERIC(10, 2) NOT NULL,
+        amount INT NOT NULL,
         date_transaction TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
         FOREIGN KEY (payer) REFERENCES users(id),
         FOREIGN KEY (payee) REFERENCES users(id)
